[Refactor] Theme 위치 변경
