+++ feature/sample/src/main/java/team/ppac/sample/SampleActivity.kt
@@ -5,7 +5,7 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.hilt.navigation.compose.hiltViewModel
 import dagger.hilt.android.AndroidEntryPoint
-import team.ppac.designsystem.foundation.FarmemeTheme
+import team.ppac.designsystem.FarmemeTheme
 
 @AndroidEntryPoint
 class SampleActivity : ComponentActivity() {
